Rewrote SSHImporter code to better convert from SSHConfig file to Tabby SSHProfiles.
