+++ tabby-electron/package.json
@@ -24,6 +24,7 @@
   "devDependencies": {
     "electron-promise-ipc": "^2.2.4",
     "ps-node": "^0.1.6",
+    "ssh-config": "^5.0.0",
     "tmp-promise": "^3.0.2",
     "which": "^3.0.0",
     "winston": "^3.3.3"
